Refactor return & replace internal components structure
