--- src/orders/components/OrderReturnPage/OrderReturnRefundItemsCard/index.tsx
@@ -1,2 +1,0 @@
-export * from "./ReturnItemsCard";
-export { default } from "./ReturnItemsCard";
