+++ src/orders/components/OrderReturnPage/components/ItemsCard/index.ts
@@ -1,0 +1,1 @@
+export * from "./ItemsCard";
