+++ src/orders/components/OrderReturnPage/components/ItemsCard/messages.ts
@@ -1,0 +1,18 @@
+export const itemsCardMessages = defineMessages({
+  improperValue: {
+    id: "xoyCZ/",
+    defaultMessage: "Improper value",
+    description: "error message",
+  },
+
+  titleFulfilled: {
+    id: "NxRsHQ",
+    defaultMessage: "Fulfillment - #{fulfilmentId}",
+    description: "section header",
+  },
+  titleUnfulfilled: {
+    id: "BkFke9",
+    defaultMessage: "Unfulfilled Items",
+    description: "section header",
+  },
+});
