+++ src/orders/components/OrderReturnPage/components/ProductErrorCell/ProductErrorCell.tsx
@@ -1,0 +1,55 @@
+import { Popper, TableCell, Typography } from "@material-ui/core";
+import ErrorExclamationCircleIcon from "@saleor/icons/ErrorExclamationCircle";
+import React, { useState } from "react";
+import { useIntl } from "react-intl";
+
+import { productErrorCellMessages } from "./messages";
+import { useProductErrorCellStyles } from "./styles";
+
+interface ProductErrorCellProps {
+  hasVariant: boolean;
+}
+
+export const ProductErrorCell: React.FC<ProductErrorCellProps> = ({
+  hasVariant,
+}) => {
+  const classes = useProductErrorCellStyles({});
+  const intl = useIntl();
+  const popperAnchorRef = React.useRef<HTMLButtonElement | null>(null);
+
+  const [showErrorBox, setShowErrorBox] = useState<boolean>(false);
+
+  if (hasVariant) {
+    return <TableCell />;
+  }
+
+  return (
+    <TableCell
+      align="right"
+      className={classes.container}
+      ref={popperAnchorRef}
+    >
+      <div
+        className={classes.titleContainer}
+        onMouseEnter={() => setShowErrorBox(true)}
+        onMouseLeave={() => setShowErrorBox(false)}
+      >
+        <Typography className={classes.errorTextHighlighted}>
+          {intl.formatMessage(productErrorCellMessages.title)}
+        </Typography>
+        <ErrorExclamationCircleIcon />
+      </div>
+      <Popper
+        placement="bottom-end"
+        open={showErrorBox}
+        anchorEl={popperAnchorRef.current}
+      >
+        <div className={classes.errorBox}>
+          <Typography className={classes.errorText}>
+            {intl.formatMessage(productErrorCellMessages.description)}
+          </Typography>
+        </div>
+      </Popper>
+    </TableCell>
+  );
+};
