+++ src/orders/components/OrderReturnPage/components/ProductErrorCell/messages.ts
@@ -1,0 +1,15 @@
+import { defineMessages } from "react-intl";
+
+export const productErrorCellMessages = defineMessages({
+  description: {
+    id: "RlbhwF",
+    defaultMessage:
+      "This product is no longer in database so it can’t be replaced, nor returned",
+    description: "product no longer exists error description",
+  },
+  title: {
+    id: "p4zuQp",
+    defaultMessage: "Product no longer exists",
+    description: "product no longer exists error title",
+  },
+});
