+++ src/orders/components/OrderReturnPage/components/index.ts
@@ -1,0 +1,3 @@
+export * from "./ItemsCard";
+export * from "./ProductErrorCell";
+export * from "./MaximalButton";
